feat(export): Saving serialized event data :D
Fixed the types a bit to use the shorter OriginalRoomMessageEvent alias,
and added `split_serialized_deserialized`, which runs over a list of
events returning their JSON data and the SDK type for those events.

The raw events get stored in `[user]/events/[user] - events.ron`, and are
equivalent to Element's "View source" function.

Only decrypted or plaintext events are stored.

Signed-off-by: Kirin Etheridge <revelation@krxt.dev>