feat(client): Multiple key files
Not the cleanest, but allows for adding multiple key files.

Also added a 2 sec `sleep` for media downloads, so 4 concurrent
downloads, but with a timeout.

This whole mechanism may be removed in the future as homeservers
*should* send a proper response about rate limiting which could be
handled (?), but on Dendrite it seems to be broken and I cannot test it
until I switch to an actively maintained server.

Signed-off-by: Kirin Etheridge <revelation@krxt.dev>