feat: New event processing, TL avatars, upd deps
This should've been a split commit, but oh well.

Added a more structured system for event processing inside
`utils/media`, which also allows for downloading historical avatars from
the timeline, and possibly more in the future. TL avatars are a bit odd
and don't contain MIME types or a filename. The SDK does provide a HTTP
`Response` type containing headers with those, but I really wanted to
stick with `get_media_file` (which internally uses that Response but only
takes the file from it), so I added `mimetype-detector` as a dependency.
See `download_media` in src/utils/media/mod.rs for more info.

Also updated dependencies, which broke the background sync task on the
default `recursion_limit`. Reading SDK code comments and issues, that
seems to be due to checking `Sync` implementation for all nested types
which with async + SDK types gets... big. As a workaround the tokio::select
was removed and syncing now happens through `sync_with_callback` doing
exactly the same thing, just with one level of `async {}` removed.

Lastly, added a quick debug flag, which will be the only CLI flag
considering this will be moved to a GUI (soon, i swear ,,,).
Use `-d` or `--debug`, no value defaults to Level::INFO.

Fixes #1, #19

Signed-off-by: Kirin Etheridge <revelation@krxt.dev>